standardize connector layout for clickhouse
